Rename block components to include block in the name
