+++ src/components/Blocks/CaseListBlock/CaseListBlock.query.ts
@@ -1,0 +1,32 @@
+import { graphql } from "~/utils/graphql";
+import { LinkFragment } from "~/components/Core/AppLink/AppLink.query";
+
+export const CaseListBlockFragment = graphql(
+  `
+    fragment CaseListBlockFragment on SectionCaseListRecord {
+      id
+      title
+      columns
+      cases {
+        title
+        slug
+        caseTeaser {
+          title
+          image {
+            url
+            alt
+            width
+            height
+          }
+        }
+      }
+      text {
+        value
+        blocks {
+          ...LinkFragment
+        }
+      }
+    }
+  `,
+  [LinkFragment],
+);
